requirements.txt
scikit-image
numpy
scipy
matplotlib
python-telegram-bot